Converted CRA to TypeScript. Installed dependencies and made any necessary changes to prevent current type errors.
