+++ client/package.json
@@ -8,11 +8,17 @@
 		"@testing-library/jest-dom": "^5.11.4",
 		"@testing-library/react": "^11.1.0",
 		"@testing-library/user-event": "^12.1.10",
+		"@types/jest": "^26.0.23",
+		"@types/node": "^15.6.0",
+		"@types/react": "^17.0.6",
+		"@types/react-dom": "^17.0.5",
+		"@types/react-router-dom": "^5.1.7",
 		"jest": "26.6.0",
 		"react": "^17.0.2",
 		"react-dom": "^17.0.2",
 		"react-router-dom": "^5.2.0",
 		"react-scripts": "4.0.3",
+		"typescript": "^4.2.4",
 		"web-vitals": "^1.0.1"
 	},
 	"proxy": "http://localhost:3001",
