--- client/src/AboutUs.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-
-function AboutUs() {
-	return (
-		<div>
-      AboutUs
-		</div>
-	);
-}
-
-export default AboutUs;
