+++ client/src/AboutUs.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+function AboutUs() {
+    return <div>AboutUs</div>;
+}
+
+export default AboutUs;
