+++ client/src/App.tsx
@@ -1,0 +1,27 @@
+import * as React from 'react';
+import { BrowserRouter } from 'react-router-dom';
+import { ThemeProvider, useTheme } from '@material-ui/core/styles';
+
+import Footer from './Footer';
+import Header from './Header';
+import Routes from './Routes';
+
+import './App.css';
+
+function App(): JSX.Element {
+    const theme = useTheme();
+
+    return (
+        <ThemeProvider theme={theme}>
+            <BrowserRouter>
+                <div className="App">
+                    <Header />
+                    <Routes />
+                    <Footer />
+                </div>
+            </BrowserRouter>
+        </ThemeProvider>
+    );
+}
+
+export default App;
