--- client/src/ContactUs.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-
-function ContactUs() {
-	return (
-		<div>
-      ContactUs
-		</div>
-	);
-}
-
-export default ContactUs;
