+++ client/src/ContactUs.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+function ContactUs() {
+    return <div>ContactUs</div>;
+}
+
+export default ContactUs;
