+++ client/src/CookiePolicy.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 
 function CookiePolicy() {
     return <div>CookiePolicy</div>;
