--- client/src/HowItWorks.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-
-function HowItWorks() {
-	return (
-		<div>
-      HowItWorks
-		</div>
-	);
-}
-
-export default HowItWorks;
