+++ client/src/HowItWorks.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+function HowItWorks() {
+    return <div>HowItWorks</div>;
+}
+
+export default HowItWorks;
