--- client/src/LibraryAndForum.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-
-function LibraryAndForum() {
-	return (
-		<div>
-      LibraryAndForum
-		</div>
-	);
-}
-
-export default LibraryAndForum;
