+++ client/src/LibraryAndForum.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+function LibraryAndForum() {
+    return <div>LibraryAndForum</div>;
+}
+
+export default LibraryAndForum;
