+++ client/src/OurStory.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 
 function OurStory() {
     return <div>OurStory</div>;
