+++ client/src/PrivacyPolicy.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 
 function PrivacyPolicy() {
     return <div>PrivacyPolicy</div>;
