+++ client/src/Routes.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Login from './Login';
 import Home from './Home';
