+++ client/src/Signup.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react'
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 import {
